Mudança de nomeclatura para se adequar ao Diagrama Relacional
