+++ app/controllers/home_controller.rb
@@ -4,7 +4,7 @@ class HomeController < ApplicationController
   def index
   	@categories = MealCategory.all
   	@meals = Meal.all
-  	@cart_item = CartItem.new
+  	@order_has_meal = OrderHasMeal.new
   end
 
 end
