+++ app/helpers/application_helper.rb
@@ -1,6 +1,6 @@
 module ApplicationHelper
 	def current_cart
-		Cart.find(session[:current_cart])
+		Order.find(session[:current_cart])
 	rescue ActiveRecord::RecordNotFound
 		render text: "Carrinho não existe."
 	end
