--- app/helpers/cart_items_helper.rb
@@ -1,2 +1,0 @@
-module CartItemsHelper
-end
