+++ app/helpers/order_has_meals_helper.rb
@@ -1,0 +1,2 @@
+module OrderHasMealsHelper
+end
