+++ app/models/Order_has_Meal.rb
@@ -1,5 +1,5 @@
-class CartItem < ApplicationRecord
-	belongs_to :cart
+class OrderHasMeal < ApplicationRecord
+	belongs_to :order
 	belongs_to :meal
 
 	validates :quantity, presence: true, numericality: { only_integer: true, greater_than: 0 }
