--- app/models/order.rb
@@ -1,5 +1,0 @@
-class Order < ApplicationRecord
-	has_many :meals
-	belongs_to :cart
-	
-end
