+++ app/views/home/index.html.erb
@@ -25,7 +25,7 @@
 			            	<p class="price"><%= number_to_currency(m.price, :unit => "R$ ", :separator => ",", :delimiter => ".") %> por porção</p>
 			            	<p class="description">Quantidade</p>
 
-			            	<%= form_for @cart_item do |f| %>
+			            	<%= form_for @order_has_meal do |f| %>
 			            		<%= f.select :quantity, [["1", 1], ["2", 2], ["3", 3], ["4", 4], ["5", 5], ["6", 6], ["7", 7], ["8", 8]], {}, {class: "form-control"} %>
 			            		<%= f.hidden_field :meal_id, value: m.id %>
