+++ db/migrate/20171219153020_create_carts.rb
@@ -1,6 +1,6 @@
 class CreateCarts < ActiveRecord::Migration[5.1]
   def change
-    create_table :carts do |t|
+    create_table :Orders do |t|
 
       t.timestamps
     end
