+++ db/migrate/20171219154254_add_quantity_to_cart.rb
@@ -1,5 +1,5 @@
 class AddQuantityToCart < ActiveRecord::Migration[5.1]
   def change
-  	add_column :carts, :quantity, :integer
+  	add_column :Orders, :quantity, :integer
   end
 end
