+++ db/migrate/20171219155140_create_cart_items.rb
@@ -1,6 +1,6 @@
 class CreateCartItems < ActiveRecord::Migration[5.1]
 	def self.up
-		create_table :cart_items do |t|
+		create_table :OrderHasMeals do |t|
 			t.column :meal_id, :integer
 			t.column :cart_id, :integer
 			t.column :price, :float
